requirements.txt
Package                      Version
---------------------------- -----------
absl-py                      2.3.1
astunparse                   1.6.3
attrs                        25.3.0
certifi                      2025.8.3
cffi                         1.17.1
charset-normalizer           3.4.3
contourpy                    1.3.3
cycler                       0.12.1
filelock                     3.19.1
flatbuffers                  25.2.10
fonttools                    4.59.2
fsspec                       2025.9.0
gast                         0.6.0
google-pasta                 0.2.0
grpcio                       1.74.0
h5py                         3.14.0
idna                         3.10
jax                          0.7.1
jaxlib                       0.7.1
Jinja2                       3.1.6
keras                        3.11.3
kiwisolver                   1.4.9
libclang                     18.1.1
Markdown                     3.9
markdown-it-py               4.0.0
MarkupSafe                   3.0.2
matplotlib                   3.10.6
mdurl                        0.1.2
mediapipe                    0.10.21
ml_dtypes                    0.5.3
mpmath                       1.3.0
namex                        0.1.0
networkx                     3.5
numpy                        1.26.3
nvidia-cublas-cu12           12.8.4.1
nvidia-cuda-cupti-cu12       12.8.90
nvidia-cuda-nvrtc-cu12       12.8.93
nvidia-cuda-runtime-cu12     12.8.90
nvidia-cudnn-cu12            9.10.2.21
nvidia-cufft-cu12            11.3.3.83
nvidia-cufile-cu12           1.13.1.3
nvidia-curand-cu12           10.3.9.90
nvidia-cusolver-cu12         11.7.3.90
nvidia-cusparse-cu12         12.5.8.93
nvidia-cusparselt-cu12       0.7.1
nvidia-nccl-cu12             2.27.3
nvidia-nvjitlink-cu12        12.8.93
nvidia-nvtx-cu12             12.8.90
opencv-contrib-python        4.8.0.76
opt_einsum                   3.4.0
optree                       0.17.0
packaging                    25.0
pillow                       11.3.0
pip                          25.2
polars                       1.33.0
protobuf                     4.25.8
psutil                       7.0.0
py-cpuinfo                   9.0.0
pycparser                    2.22
Pygments                     2.19.2
pyparsing                    3.2.3
python-dateutil              2.9.0.post0
PyYAML                       6.0.2
requests                     2.32.5
rich                         14.1.0
scipy                        1.16.1
sentencepiece                0.2.1
setuptools                   80.9.0
six                          1.17.0
sounddevice                  0.5.2
sympy                        1.14.0
tensorboard                  2.19.0
tensorboard-data-server      0.7.2
tensorflow                   2.19.0
tensorflow-io-gcs-filesystem 0.37.1
termcolor                    3.1.0
torch                        2.8.0
torchvision                  0.23.0
triton                       3.4.0
typing_extensions            4.15.0
ultralytics                  8.3.194
ultralytics-thop             2.0.17
urllib3                      2.5.0
Werkzeug                     3.1.3
wheel                        0.45.1
wrapt                        1.17.3
